requirements.txt
asgiref==3.11.0
black==26.1.0
click==8.3.1
Django==5.2.10
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
flake8==7.3.0
iniconfig==2.3.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==1.0.3
platformdirs==4.5.1
pluggy==1.6.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-django==4.11.1
pytokens==0.4.0
sqlparse==0.5.5
